# column-major access and other perf improvements for generic triangular solves