This pull request addresses #14471 by replacing the `naivesub!` methods in base/linalg/triangular.jl with column-major-access versions. The new methods also benefit from `@inbounds` decoration and direct indexing of the object underlying the triangular object; each of these two modifications improves performance significantly on top of the change to column-major access. Benchmark code:
```julia
using Benchmarks

m = 5000;
bo = rand(m);
b = bo;
A = eye(m) + randn(m, m)/(2m);
ltA = LowerTriangular(A);
utA = UpperTriangular(A);
ultA = Base.LinAlg.UnitLowerTriangular(A);
uutA = Base.LinAlg.UnitUpperTriangular(A);

# Benchmark
for (matstr, mat) in (
    ("LowerTriangular", ltA), ("UnitLowerTriangular", ultA),
    ("UpperTriangular", utA), ("UnitUpperTriangular", uutA) )
    copy!(b, bo)
    println(matstr)
    show(@benchmark Base.LinAlg.naivesub!(mat, b))
    println()
end
```
On master these benchmarks yield:
```julia
LowerTriangular
================ Benchmark Results ========================
     Time per evaluation: 143.57 ms [141.11 ms, 146.02 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 67
   Number of evaluations: 67
 Time spent benchmarking: 9.78 s

UnitLowerTriangular
================ Benchmark Results ========================
     Time per evaluation: 167.37 ms [164.84 ms, 169.91 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 57
   Number of evaluations: 57
 Time spent benchmarking: 9.73 s

UpperTriangular
================ Benchmark Results ========================
     Time per evaluation: 142.41 ms [140.09 ms, 144.72 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 69
   Number of evaluations: 69
 Time spent benchmarking: 10.00 s

UnitUpperTriangular
================ Benchmark Results ========================
     Time per evaluation: 166.16 ms [163.87 ms, 168.46 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 57
   Number of evaluations: 57
 Time spent benchmarking: 9.65 s
```
On this PR's branch these benchmarks yield:
```julia
LowerTriangular
================ Benchmark Results ========================
  Time per evaluation: 9.13 ms [8.57 ms, 9.68 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
     Memory allocated: 0.00 bytes
Number of allocations: 0 allocations
    Number of samples: 100
Number of evaluations: 100
Time spent benchmarking: 1.00 s

UnitLowerTriangular
================ Benchmark Results ========================
  Time per evaluation: 9.19 ms [8.60 ms, 9.77 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
     Memory allocated: 0.00 bytes
Number of allocations: 0 allocations
    Number of samples: 100
Number of evaluations: 100
Time spent benchmarking: 0.96 s

UpperTriangular
================ Benchmark Results ========================
  Time per evaluation: 10.64 ms [10.03 ms, 11.26 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
     Memory allocated: 0.00 bytes
Number of allocations: 0 allocations
    Number of samples: 100
Number of evaluations: 100
Time spent benchmarking: 1.12 s

UnitUpperTriangular
================ Benchmark Results ========================
  Time per evaluation: 10.65 ms [10.05 ms, 11.24 ms]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
     Memory allocated: 0.00 bytes
Number of allocations: 0 allocations
    Number of samples: 100
Number of evaluations: 100
Time spent benchmarking: 1.11 s
```

Point of curiosity: In the methods for upper triangular matrices, I expected backward iteration within columns (`i = j-1:-1:n`, with `j` the column index and `i` the row index) to perform better than forward iteration within columns (`i = 1:j-1`) given the immediately preceding accesses of `x[j]`, `b[j]`, and `A[j,j]`. But to my surprise the forward iteration is faster: Forward iteration in the upper triangular methods yields performance parity with the lower triangular methods. The native code looks essentially identical. Something with cache behavior? Thoughts? I left the backward iteration in for now, suspecting that this performance observation is hardware-specific and forward iteration might be surprising to readers.

I was surprised by how much direct indexing (`A.data`) impacted performance, particularly in the unit triangular case. I wonder whether the overhead of indexing through the triangular wrapper could be reduced. Thoughts? I will play with this a bit and open an issue if I find anything interesting.

Manually hoisting the `A.data` reference out of the loops did not impact performance significantly.

These and related methods in `triangular.jl` cry out for unification. I will throw together a concept PR.