@KristofferC 2015-12-23
The Triangular types are immutable and in those cases LLVM is usually able to hoist the load.