@Sacha0 2015-12-23
The diff was unsightly due to accidental reordering of the method definitions. Now more sane.