@andreasnoack 2015-12-28
This great although this methods shouldn't be called for most element types for which memory access is the bottleneck (since they should use BLAS).

Usually, we don't make comments in the source like the ones you have added, so I'd suggest that they are removed. You can put them in a comment in this issue. In that way, we can find the conclusions from your benchmarking again when we need them.