@tkelman 2015-12-28
Lack of comments in the source is a bug, not a feature.