@nalimilan 2015-12-28
> Lack of comments in the source is a bug, not a feature.

+100 I think we should push for more comments from all developers.