@Sacha0 2015-12-28
> this methods shouldn't be called for most element types for which memory access is the bottleneck (since they should use BLAS).

Absolutely. A more realistic test case for such methods would be great. A performant, light-weight extended-precision type like double-double came to mind. Can you think of others?