@Sacha0 2016-01-10
Consensus on the comments? Should they stay or go? Is this otherwise in shape to merge? Thanks!