@andreasnoack 2016-01-10
Sorry that this stalled. The level of detail in the comments here is higher than I find suitable, e.g. I think that the observations you made could easily change with compiler changes, but this is really a minor issue and @tkelman and @nalimilan are in favor so I'll merge as it is. Thanks for the contribution and patience.