@Sacha0 2016-01-10
> Sorry that this stalled.

No need for apology! Only so many hours in the day, and it was a holiday besides :).

> Thanks for the contribution and patience.

Thanks for the review and merge! Best,